update embed color and guild creation

+++ commands/suggestion.js
@@ -3,7 +3,10 @@ module.exports = {
     name: "suggest",
     description: `Suggest a feature for Sir Countalot`,
     permissions: `Be TheWaildug (as of right now)`,
-    execute(message,args,trello){
-        return message.delete();
+    execute(message,args){
+        if(message.member.id != "432345618028036097"){
+            return message.delete();
+        }
+        
     }
 }
